Add recipe image, steps, and improve styling.
This commit adds the recipe image to the list view with a forced size.
It also adds the IndividualRecipeCard, which is what the individual
drink/recipe view uses. This will allow for better separation, for
example showing steps and an "order" button for the individual view.

Included some stylistic changes to CSS, including spacing between
ingredients and steps for the individual recipe view - but it is a bit
too spaced.